# Optimize _looksLikeClassName() in AOT builds.